bqapi : remove redundant code
After discussing with Connor, we concluded that
bqnode.py is copy of bqclass.py (redundant code)
and bqnode.py is not even used as seen in types.py.

Not sure about goal to create the same as the changes between
two files includes some TAG variable and passing xml
structure instead of object to class BQNode.

So removing for now to clean the codebase !

Signed-off-by: charvi-077 <charvi077@gmail.com>